+++ components/navbar.html
@@ -4,8 +4,9 @@
 
         <div class="d-flex justify-content-between gap-5">
             <div class="d-flex w-25 d-md-none justify-content-center gap-4 align-items-center">
-                <a class="nav-link" href="#" title="Cart">
+                <a class="nav-link position-relative" href="#" title="Cart" type="button"  data-bs-toggle="offcanvas" data-bs-target="#cartCanvas" aria-controls="cartCanvas">
                     <i class="nav-icon fas fa-cart-shopping fs-5"></i>
+                    <span class="position-absolute top-0 start-0 translate-middle badge rounded-pill bg-danger">2</span>
                 </a>
                 <a class="nav-link" href="#" title="Wishlist">
                     <i class="nav-icon fa-regular fa-heart fs-5"></i>
@@ -57,8 +58,9 @@
         </div>
 
         <div class="d-none d-md-flex align-items-center justify-content-center gap-4">
-            <a class="nav-link" href="#" title="Cart">
+            <a class="nav-link position-relative" href="#" title="Cart" type="button"  data-bs-toggle="offcanvas" data-bs-target="#cartCanvas" aria-controls="cartCanvas">
                 <i class="nav-icon fas fa-cart-shopping fs-5"></i>
+                <span class="position-absolute top-0 start-0 translate-middle badge rounded-pill bg-danger">2</span>
             </a>
             <a class="nav-link" href="#" title="Wishlist">
                 <i class="nav-icon fa-regular fa-heart fs-5"></i>
